+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Security`
 
-## Fixed
+
+## Unreleased
+
+- SC-6239 - adds lernstore permmission check
+- SC-6239 - permmission check now supports also NOT operator (besides OR, AND)
+
+### Fixed
 
 - SC-6357: Fixed the bouble loading issue in admin tables when changing the number of elements to display